requirements.txt
pytest == 7.1.2
selenium == 4.0.0
pytest-html == 3.1.1
pytest-xdist == 2.5.0
openpyxl == 3.0.9
allure-pytest == 2.9.45
webdriver-manager == 3.5.4
Pillow == 9.1.0
pymongo == 4.1.1
mysql-connector-python == 8.0.29
protobuf == 3.20.1
cryptocode == 0.1

